Organize format code in a subdirectory
