+++ src/common/format/template.js
